Fixed #11 - Allow editing of displayed names for each caller.
Client has edit box instead of div for caller edit mode.
Server creates a jcadmin.json database file to hold names for each caller
regardless of how that caller is added/removed from whitelist and blacklist.
But we keep names up to date as comments in the whitelist and blacklist.