Fixed #21 - Enumerate all dates/times of calls received by selected caller.
The details page for a phone number now puts a table of all the dates
and times that calls have been received from that phone number.
If no calls have been received from that number, we put explanatory text
there.

Removed row for call count, because it is redundant with the new list.

The method /api/caller/:phonenumber now returns complete list of all
date/time for the given phone number, in reverse chronological order.